// src/model/Theme.ts
export interface ITheme {
    bodyBackground: Number[];

    headerBackground: Number[];
    headerTitleColor: Number[];

    inputNewTaskBackground: Number[];
    buttonAddNewTaskBackground: Number[];
    
    primariColor: Number[];

    primaryTextColor: Number[];
    secondaryTextColor: Number[];

    checkmarkTaskBackground: Number[];
    checkmarkTaskMarkBackground: Number[];
    checkmarkTaskMarkBorder: Number[];
    checkmarkTaskTextColor: Number[];

    checkmarkedTaskBackground: Number[];
    checkmarkedTaskMarkBackground: Number[];
    checkmarkedTaskMarkBorder: Number[];
    checkmarkedTaskTextColor: Number[];
}